chore: split user views into multiple files
